packaging/adk_stop_loop_agent/pyproject.toml
[project]
name = "adk_stop_loop_agent"
version = "0.0.1"
dependencies = ["google-adk"]
